Refactor doltlite merge into rows, schema, and catalog modules
Split the ~3.5k-line doltlite_merge.c along natural seams and decompose
mergeCatalogPass1 so the highest-risk path is reviewable without behavior
changes.

- doltlite_merge_rows.c: index-key helpers and three-way row/cell merge
- doltlite_merge_schema.c: CREATE TABLE parse, schema IR, column merge
- doltlite_merge.c: catalog orchestration with MergePass1Ctx helpers
- Wire main.mk/amalgamation, layer lint (2.5k line cap), AGENTS.md notes

Verified: merge, index-conflict, ignore-corners, and cherry-pick shell suites.
